+++ js_api/src/service/gov.ts
@@ -213,7 +213,7 @@ async function getTreasuryTips(api: ApiPromise) {
  * make an extrinsic of treasury proposal submission for council member.
  */
 async function makeTreasuryProposalSubmission(api: ApiPromise, id: any, isReject: boolean): Promise<SubmittableExtrinsic<"promise">> {
-  const members = await api.query.electionsPhragmen.members<any[]>();
+  const members = await (api.query.electionsPhragmen || api.query.elections || api.query.phragmenElection).members<any[]>();
   const councilThreshold = Math.ceil(members.length * 0.6);
   const proposal = isReject ? api.tx.treasury.rejectProposal(id) : api.tx.treasury.approveProposal(id);
   return api.tx.council.propose(councilThreshold, proposal, proposal.length);