+++ lib/service/account.dart
@@ -9,28 +9,35 @@ class ServiceAccount {
 
   /// encode addresses to publicKeys
   Future<Map> encodeAddress(List<String> pubKeys, ss58List) async {
-    Map res = await serviceRoot.webView.evalJavascript(
+    final Map res = await serviceRoot.webView.evalJavascript(
         'account.encodeAddress(${jsonEncode(pubKeys)}, ${jsonEncode(ss58List)})');
     return res;
   }
 
   /// decode addresses to publicKeys
   Future<Map> decodeAddress(List<String> addresses) async {
-    Map res = await serviceRoot.webView
+    final Map res = await serviceRoot.webView
         .evalJavascript('account.decodeAddress(${jsonEncode(addresses)})');
+    return res;
+  }
+
+  /// check address matches ss58Format
+  Future<bool> checkAddressFormat(String address, int ss58) async {
+    final bool res = await serviceRoot.webView
+        .evalJavascript('account.checkAddressFormat("$address", $ss58)');
     return res;
   }
 
   /// query balance
   Future<Map> queryBalance(String address) async {
-    final res = await serviceRoot.webView
+    final Map res = await serviceRoot.webView
         .evalJavascript('account.getBalance(api, "$address")');
     return res;
   }
 
   /// Get on-chain account info of addresses
   Future<List> queryIndexInfo(List addresses) async {
-    var res = await serviceRoot.webView.evalJavascript(
+    final List res = await serviceRoot.webView.evalJavascript(
         'account.getAccountIndex(api, ${jsonEncode(addresses)})');
     return res;
   }
@@ -43,13 +50,13 @@ class ServiceAccount {
   }
 
   Future<List> getPubKeyIcons(List<String> keys) async {
-    List res = await serviceRoot.webView
+    final List res = await serviceRoot.webView
         .evalJavascript('account.genPubKeyIcons(${jsonEncode(keys)})');
     return res;
   }
 
   Future<List> getAddressIcons(List addresses) async {
-    List res = await serviceRoot.webView
+    final List res = await serviceRoot.webView
         .evalJavascript('account.genIcons(${jsonEncode(addresses)})');
     return res;
   }